+++ posthog-web/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.4.0 - 2023-04-20
+
+1. Fixes a race condition that could occur when initialising PostHog
+2. Fixes an issue where feature flags would not be reloaded after a reset
+
 # 2.3.0 - 2023-04-19
 
 1. Some small fixes to incorrect types